Refactor example js
