+++ examples/static/src/js/icons.js
@@ -1,0 +1,15 @@
+(function () {
+  'use strict';
+
+  var icons = [
+    'menu',
+    'chevron-left',
+    'chevron-right',
+    'chevron-up',
+    'chevron-down'
+  ];
+
+  icons.sort();
+
+  module.exports = icons;
+})();
